refactor cron auth helper
