+++ app/api/cron/update-digest/route.ts
@@ -1,15 +1,7 @@
 import { NextResponse } from 'next/server'
+import { requireCronAuth } from '@/lib/api/cron-auth'
 import { listUsersWithPendingUpdates } from '@/lib/memory/updates'
 import { summarizePendingUpdatesForUser, type UpdateSummarizerResult } from '@/lib/memory/update-runner'
-
-function requireCronAuth(req: Request): boolean {
-  const secret = process.env.CRON_SECRET
-  if (!secret) {
-    return process.env.NODE_ENV !== 'production'
-  }
-  const authHeader = req.headers.get('authorization')
-  return authHeader === `Bearer ${secret}`
-}
 
 async function runUpdateDigest(): Promise<Response> {
   const users = await listUsersWithPendingUpdates()
